modify invoices styles, lift states up
