+++ src/pages/Invoices/InvoicesDash.jsx
@@ -1,7 +1,14 @@
 import React, { useState, useEffect, useMemo } from 'react';
 import { connect } from 'react-redux';
 import { useHistory } from 'react-router-dom';
-import { Button, makeStyles, Tab, Tabs } from '@material-ui/core';
+import {
+    Button,
+    makeStyles,
+    Tab,
+    Tabs,
+    FormControlLabel,
+    Switch,
+} from '@material-ui/core';
 import {
     getInvoicesAct,
     deleteInvoicesAct,
@@ -15,6 +22,24 @@ import InvoicesList from './InvoicesList';
 const useStyles = makeStyles({
     button: {
         margin: '10px',
+    },
+    container: {
+        margin: '30px auto',
+    },
+    searchBar: {
+        width: '30vw',
+        textAlign: 'center',
+    },
+    tabHeader: {
+        width: '40vw',
+    },
+    options: {
+        width: '80vw',
+        display: 'flex',
+        justifyContent: 'space-between',
+    },
+    list: {
+        width: '80vw',
     },
 });
 
@@ -36,6 +61,7 @@ const InvoicesDash = props => {
     const [selected, setSelected] = useState([]);
     const [isDeleted, setIsDeleted] = useState(false);
     const [deleteAlertOpen, setDeleteAlertOpen] = useState(false);
+    const [dense, setDense] = useState(false);
     const classes = useStyles();
     const businessesById = useMemo(() => arrToObj(businesses, 'id'), [
         businesses,
@@ -86,6 +112,10 @@ const InvoicesDash = props => {
         setTab(newValue);
     };
 
+    const handleChangeDense = event => {
+        setDense(event.target.checked);
+    };
+
     const deleteInvoices = () => {
         if (!isLoggedIn) {
             let deleteInvoicesList = Object.assign(invoicesList);
@@ -110,7 +140,7 @@ const InvoicesDash = props => {
     };
 
     return (
-        <div>
+        <div className={classes.container}>
             {deleteAlertOpen && (
                 <DeleteAlert
                     deleteAlertOpen={deleteAlertOpen}
@@ -120,44 +150,59 @@ const InvoicesDash = props => {
                     status={status}
                 />
             )}
-            <div>Search</div>
-            <div>
+            <div className={`${classes.container} ${classes.searchBar}`}>
+                Search
+            </div>
+            <div className={`${classes.container} ${classes.tabHeader}`}>
                 <Tabs indicatorColor="primary" value={tab} onChange={tabChange}>
                     <Tab label="All" value="all" />
                     <Tab label="Outstanding" value="outstanding" />
                     <Tab label="Paid" value="paid" />
                 </Tabs>
-                <Button
-                    variant="contained"
-                    className={classes.button}
-                    onClick={() => {
-                        history.push(`${history.location.pathname}/new`);
-                    }}
-                >
-                    New
-                </Button>
-                {selected.length > 0 && (
-                    <Button variant="contained" className={classes.button}>
-                        Send
-                    </Button>
-                )}
-                {selected.length > 0 && (
+            </div>
+            <div className={`${classes.container} ${classes.options}`}>
+                <div>
                     <Button
                         variant="contained"
                         className={classes.button}
                         onClick={() => {
-                            setDeleteAlertOpen(true);
+                            history.push(`${history.location.pathname}/new`);
                         }}
                     >
-                        DELETE
+                        New
                     </Button>
-                )}
+                    {selected.length > 0 && (
+                        <Button variant="contained" className={classes.button}>
+                            Send
+                        </Button>
+                    )}
+                    {selected.length > 0 && (
+                        <Button
+                            variant="contained"
+                            className={classes.button}
+                            onClick={() => {
+                                setDeleteAlertOpen(true);
+                            }}
+                        >
+                            DELETE
+                        </Button>
+                    )}
+                </div>
+                <FormControlLabel
+                    control={
+                        <Switch checked={dense} onChange={handleChangeDense} />
+                    }
+                    label="Dense padding"
+                />
             </div>
-            <InvoicesList
-                invoicesList={invoicesList}
-                selected={selected}
-                setSelected={setSelected}
-            />
+            <div className={`${classes.container} ${classes.list}`}>
+                <InvoicesList
+                    invoicesList={invoicesList}
+                    selected={selected}
+                    setSelected={setSelected}
+                    dense={dense}
+                />
+            </div>
         </div>
     );
 };
